Check valid flags in global tropo model, not just values.
The constructors do not initialize the double values but do set the valid flags
to false. Checking only height != ht, for example, without checking if height is
valid to begin with can cause the if statement to fail if whatever value
lingering in height's memory happens to match ht. In that case, validHeight will
never be set and the model will never be initialized.

In particular, this can happen if the model instance is allocated on the stack
in a loop, in which case it will end up in the same memory location on each
iteration but the constructor will reset the valid* flags each time.